Refactored Controllers directory
